+++ examples/apps/auth-sample/src/hooks/useOpenfort.tsx
@@ -3,7 +3,10 @@ import {
   RecoveryParams,
   RecoveryMethod,
   type Provider,
-  type TypedDataPayload
+  type TypedDataPayload,
+  EmbeddedAccount,
+  AccountTypeEnum,
+  ChainTypeEnum
 } from '@openfort/openfort-js';
 import axios from 'axios';
 import type React from 'react';
@@ -20,22 +23,12 @@ import { Address } from 'viem/accounts';
 import { polygonAmoy } from 'viem/chains';
 import openfort from '../utils/openfortConfig';
 
+const chainId = Number(process.env.NEXT_PUBLIC_CHAIN_ID);
+
 interface ContextType {
   state: EmbeddedState;
   getEvmProvider: () => Promise<Provider>;
-  handleRecovery: ({
-    method,
-    password,
-    chainId
-  }: {
-    method: RecoveryMethod,
-    chainId: number
-    password?: string,
-  }) => Promise<void>;
-  setWalletRecovery: (
-    recoveryMethod: RecoveryMethod,
-    recoveryPassword: string
-  ) => Promise<{ error?: Error }>;
+  handleRecovery: ({ account, recoveryParams }: { account: string, recoveryParams: RecoveryParams }) => Promise<EmbeddedAccount>;
   exportPrivateKey: () => Promise<{ data?: string; error?: Error }>;
   signMessage: (
     message: string,
@@ -49,6 +42,14 @@ interface ContextType {
   logout: () => Promise<void>;
   getEOAAddress: () => Promise<string>;
   getBalance: (address: Address, chain: Chain, provider: Provider) => Promise<bigint>;
+  account: EmbeddedAccount | null;
+  getEncryptionSession: () => Promise<string>;
+  accounts: EmbeddedAccount[];
+  isLoadingAccounts: boolean;
+  refetchAccounts: () => Promise<void>;
+  refetchAccount: () => Promise<void>;
+  createWallet: ({ recoveryParams }: { recoveryParams: RecoveryParams }) => Promise<EmbeddedAccount>;
+  setRecoveryMethod: (previousRecovery: RecoveryParams, newRecovery: RecoveryParams) => Promise<void>;
 }
 
 const OpenfortContext = createContext<ContextType | null>(null);
@@ -97,6 +98,39 @@ export const OpenfortProvider: React.FC<React.PropsWithChildren<unknown>> = ({
       if (poller.current) clearInterval(poller.current);
     };
   }, []);
+
+  // Account
+  const [account, setAccount] = useState<EmbeddedAccount | null>(null);
+  const [accounts, setAccounts] = useState<EmbeddedAccount[]>([]);
+  const [isLoadingAccounts, setIsLoadingAccounts] = useState(false);
+
+  const refetchAccount = useCallback(async () => {
+    const account = await openfort.embeddedWallet.get();
+    setAccount(account);
+  }, []);
+
+  const refetchAccounts = useCallback(async () => {
+    setIsLoadingAccounts(true);
+    const accounts = await openfort.embeddedWallet.list({
+      chainId: chainId
+    });
+    setAccounts(accounts);
+    setIsLoadingAccounts(false);
+  }, []);
+
+  useEffect(() => {
+    if (state === EmbeddedState.EMBEDDED_SIGNER_NOT_CONFIGURED) {
+      setAccount(null);
+      refetchAccounts();
+    } else if (state === EmbeddedState.READY) {
+      refetchAccount();
+      refetchAccounts();
+    } else if (state === EmbeddedState.UNAUTHENTICATED) {
+      setAccount(null);
+      setAccounts([]);
+    }
+  }, [state]);
+
 
   const getEvmProvider = useCallback(async (): Promise<Provider> => {
     const externalProvider = await openfort.embeddedWallet.getEthereumProvider({
@@ -150,51 +184,6 @@ export const OpenfortProvider: React.FC<React.PropsWithChildren<unknown>> = ({
     }
   }, []);
 
-  const setWalletRecovery = useCallback(
-    async (
-      recoveryMethod: RecoveryMethod,
-      recoveryPassword: string
-    ): Promise<{ error?: Error }> => {
-      try {
-        switch (recoveryMethod) {
-          case RecoveryMethod.AUTOMATIC:
-            await openfort.embeddedWallet.setRecoveryMethod({
-              recoveryMethod: RecoveryMethod.PASSWORD,
-              password: recoveryPassword,
-            }, {
-              recoveryMethod,
-              encryptionSession: await getEncryptionSession(),
-            });
-            break;
-          case RecoveryMethod.PASSWORD:
-            if (!recoveryPassword || recoveryPassword.length < 4) {
-              throw new Error('Password recovery must be at least 4 characters');
-            }
-            await openfort.embeddedWallet.setRecoveryMethod({
-              recoveryMethod: RecoveryMethod.AUTOMATIC,
-              encryptionSession: await getEncryptionSession(),
-            }, {
-              recoveryMethod,
-              password: recoveryPassword,
-            });
-            break;
-          default:
-            throw new Error('Invalid recovery method');
-        }
-        return {};
-      } catch (err) {
-        console.error('Error setting wallet recovery:', err);
-        return {
-          error:
-            err instanceof Error
-              ? err
-              : new Error('An error occurred setting wallet recovery'),
-        };
-      }
-    },
-    []
-  );
-
   const signTypedData = useCallback(
     async (
       domain: TypedDataPayload['domain'],
@@ -217,37 +206,14 @@ export const OpenfortProvider: React.FC<React.PropsWithChildren<unknown>> = ({
     []
   );
 
-  const handleRecovery = useCallback(async ({ method, password, chainId }: { method: RecoveryMethod, password?: string, chainId: number }) => {
-    let recoveryParams: RecoveryParams;
-    switch (method) {
-      case 'automatic':
-        recoveryParams = {
-          recoveryMethod: RecoveryMethod.AUTOMATIC,
-          encryptionSession: await getEncryptionSession()
-        };
-        break;
-      case 'password':
-        if (!password || password.length < 4) {
-          throw new Error('Password recovery must be at least 4 characters');
-        }
-        recoveryParams = {
-          recoveryMethod: RecoveryMethod.PASSWORD,
-          password,
-        };
-        break;
-      case 'passkey':
-        recoveryParams = {
-          recoveryMethod: RecoveryMethod.PASSKEY,
-        }
-        break;
-      default:
-        throw new Error('Invalid recovery method');
-    }
-
-    await openfort.embeddedWallet.configure({
-      chainId,
-      recoveryParams
+  const handleRecovery = useCallback(async ({ account, recoveryParams }: { account: string, recoveryParams: RecoveryParams }) => {
+    const response = await openfort.embeddedWallet.recover({
+      recoveryParams,
+      account,
     });
+    refetchAccounts();
+    setAccount(response);
+    return response;
   }, []);
 
   const logout = useCallback(async () => {
@@ -288,17 +254,49 @@ export const OpenfortProvider: React.FC<React.PropsWithChildren<unknown>> = ({
     }
   }, []);
 
+  const createWallet = useCallback(async ({ recoveryParams }: { recoveryParams: RecoveryParams }) => {
+    const user = await openfort.user.get()
+    if (!user) throw new Error('User not found');
+
+    if (!!user.linkedAccounts.some(la => la.email === "testing@fort.dev")) {
+      throw new Error('Test user should not create wallets.');
+    }
+
+    const response = await openfort.embeddedWallet.create({
+      accountType: AccountTypeEnum.SMART_ACCOUNT,
+      chainType: ChainTypeEnum.EVM,
+      recoveryParams,
+      chainId,
+    })
+    setAccount(response);
+    await refetchAccounts();
+    return response;
+  }, [refetchAccounts]);
+
+  const setRecoveryMethod = useCallback(async (previousRecovery: RecoveryParams, newRecovery: RecoveryParams) => {
+    await openfort.embeddedWallet.setRecoveryMethod(previousRecovery, newRecovery);
+    await refetchAccount();
+    await refetchAccounts();
+  }, [refetchAccount, refetchAccounts]);
+
   const contextValue: ContextType = {
     state,
+    getEncryptionSession,
     getEvmProvider,
     handleRecovery,
     signMessage,
-    setWalletRecovery,
     exportPrivateKey,
     signTypedData,
     logout,
     getEOAAddress,
     getBalance,
+    account,
+    refetchAccount,
+    accounts,
+    isLoadingAccounts,
+    refetchAccounts,
+    createWallet,
+    setRecoveryMethod,
   };
 
   return (