+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.8] - 2025-09-03
+### Feat
+- Wallet returns recovery method
+
+## [0.10.7] - 2025-09-03
+### Fix
+- Change account selecting during configure. Now it selects accounts by chain id if it's present
+## [0.10.6] - 2025-09-03
+### Feat
+- Remove token refreshed event and related code
+
 ## [0.10.5] - 2025-08-29
 ### Feat
 - Change API methods which configure method calls. Now it calls V2 methods